requirements.txt
cffi==1.15.1
einops==0.6.0
filelock==3.9.0
googleapis-common-protos==1.58.0
grpclib==0.4.3
h2==4.1.0
hpack==4.0.0
huggingface-hub==0.13.1
hyperframe==6.0.1
inflect==6.0.2
lazy_loader==0.1
librosa==0.10.0
msgpack==1.0.5
multidict==6.0.4
mypy-protobuf==3.4.0
numpy==1.24.3
Pillow==9.4.0
ply==3.11
pooch==1.7.0
progressbar==2.5
protobuf==4.22.1
pycparser==2.21
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rotary-embedding-torch==0.2.1
scipy==1.10.1
sip==6.7.9
soundfile==0.12.1
soxr==0.3.4
TBB==0.2
tokenizers==0.13.2
TorToiSe==2.4.2
tqdm==4.65.0
transformers==4.26.1
types-protobuf==4.23.0.1
typing_extensions==4.5.0
Unidecode==1.3.6
viam-sdk==0.2.18
